Extract common UserContent behavior
